+++ remote-communication/pkg/communication/adapters/rabbitmq/subscribe_event_handler_test.go
@@ -1,55 +1,81 @@
 package rabbitmq
 
 import (
+	"remote-communication/pkg/communication/app"
 	"testing"
 
+	"github.com/Tomofiles/skysign_cloud_v2/skysign-proto/pkg/skysign_proto"
 	"github.com/stretchr/testify/assert"
+	"github.com/stretchr/testify/mock"
+	"google.golang.org/protobuf/proto"
 )
 
 // TestSubscribeEventHandlerCommunicationIdGaveEvent .
 func TestSubscribeEventHandlerCommunicationIdGaveEvent(t *testing.T) {
 	a := assert.New(t)
 
+	service := manageCommunicationServiceMock{}
+	service.On("CreateCommunication", mock.Anything).Return(nil)
+
+	app := app.Application{
+		Services: app.Services{
+			ManageCommunication: &service,
+		},
+	}
+
 	psm := &publishHandlerMock{}
-	evt := &eventHandlerMock{}
-	SubscribeEventHandler(nil, psm, evt)
+	SubscribeEventHandler(nil, psm, app)
+
+	requestPb := &skysign_proto.CommunicationIdGaveEvent{
+		CommunicationId: string(DefaultCommunicationID),
+	}
+	requestBin, _ := proto.Marshal(requestPb)
 
 	var (
 		ExchangeName = "vehicle.communication_id_gave_event"
 		QueueName    = "communication.communication_id_gave_event"
-		EventByte    = []byte{0x20, 0x21}
 	)
 
 	for _, c := range psm.consumers {
 		if c.exchangeName == ExchangeName && c.queueName == QueueName {
-			c.handler(EventByte)
+			c.handler(requestBin)
 		}
 	}
 
-	a.Equal(evt.events1, EventByte)
-	a.Empty(evt.events2)
+	a.Equal(service.ID, string(DefaultCommunicationID))
 }
 
 // TestSubscribeEventHandlerCommunicationIdRemovedEvent .
 func TestSubscribeEventHandlerCommunicationIdRemovedEvent(t *testing.T) {
 	a := assert.New(t)
 
+	service := manageCommunicationServiceMock{}
+	service.On("DeleteCommunication", mock.Anything).Return(nil)
+
+	app := app.Application{
+		Services: app.Services{
+			ManageCommunication: &service,
+		},
+	}
+
 	psm := &publishHandlerMock{}
-	evt := &eventHandlerMock{}
-	SubscribeEventHandler(nil, psm, evt)
+	SubscribeEventHandler(nil, psm, app)
+
+	requestPb := &skysign_proto.CommunicationIdRemovedEvent{
+		CommunicationId: string(DefaultCommunicationID),
+	}
+	requestBin, _ := proto.Marshal(requestPb)
 
 	var (
 		ExchangeName = "vehicle.communication_id_removed_event"
 		QueueName    = "communication.communication_id_removed_event"
-		EventByte    = []byte{0x20, 0x21}
 	)
 
 	for _, c := range psm.consumers {
 		if c.exchangeName == ExchangeName && c.queueName == QueueName {
-			c.handler(EventByte)
+			c.handler(requestBin)
 		}
 	}
 
-	a.Empty(evt.events1)
-	a.Equal(evt.events2, EventByte)
+	a.Equal(service.ID, string(DefaultCommunicationID))
 }
